# [lte][agw] Update CPE monitoring last reported time to milliseconds