--- a/lte/cloud/go/services/subscriberdb/obsidian/models/icmp_status_swaggergen.go
+++ b/lte/cloud/go/services/subscriberdb/obsidian/models/icmp_status_swaggergen.go
@@ -17,8 +17,8 @@ import (
 // swagger:model icmp_status
 type IcmpStatus struct {
 
-	// last reported time
-	LastReportedTime float32 `json:"last_reported_time,omitempty"`
+	// Timestamp of last reported status for the subscriber in ms
+	LastReportedTime int64 `json:"last_reported_time,omitempty"`
 
 	// latency ms
 	// Required: true