--- a/lte/cloud/go/services/subscriberdb/obsidian/models/swagger.v1.yml
+++ b/lte/cloud/go/services/subscriberdb/obsidian/models/swagger.v1.yml
@@ -503,8 +503,9 @@ definitions:
         format: float
         example: 12.34
       last_reported_time:
-        type: integer
-        format: int64
+        type: number
+        format: float
+        example: 1605747300000
 
   apn_list:
     type: array