--- a/lte/cloud/go/services/subscriberdb/obsidian/models/swagger.v1.yml
+++ b/lte/cloud/go/services/subscriberdb/obsidian/models/swagger.v1.yml
@@ -503,8 +503,9 @@ definitions:
         format: float
         example: 12.34
       last_reported_time:
-        type: number
-        format: float
+        description: Timestamp of last reported status for the subscriber in ms
+        type: integer
+        format: int64
         example: 1605747300000
 
   apn_list: