--- a/lte/gateway/configs/pipelined.yml
+++ b/lte/gateway/configs/pipelined.yml
@@ -131,7 +131,7 @@ redis_enabled: false
 magma_print_grpc_payload: false
 
 # MTR iface IP
-mtr_ip: 10.0.2.10
+mtr_ip: 10.1.0.1
 mtr_interface: mtr0
 
 ###############