--- a/lte/gateway/python/magma/monitord/main.py
+++ b/lte/gateway/python/magma/monitord/main.py
@@ -20,6 +20,10 @@ import logging
 from lte.protos.mobilityd_pb2 import IPAddress
 
 
+def _get_serialized_subscriber_states(cpe_monitor: CpeMonitoringModule):
+    return serialize_subscriber_states(cpe_monitor.get_subscriber_state())
+
+
 def main():
     """ main() for monitord service"""
     manual_ping_targets = {}
@@ -28,6 +32,7 @@ def main():
     # Monitoring thread loop
     mtr_interface = load_service_config("monitord")["mtr_interface"]
 
+    # Add manual IP targets from yml file
     try:
         targets = load_service_config("monitord")["ping_targets"]
         for target, data in targets.items():
@@ -51,8 +56,7 @@ def main():
 
     # Register a callback function for GetOperationalStates
     service.register_operational_states_callback(
-        lambda: serialize_subscriber_states(
-            cpe_monitor.get_subscriber_state()))
+        _get_serialized_subscriber_states(cpe_monitor))
 
     # Run the service loop
     service.run()