<!--
    Tag your PR title with the components that it touches.
    E.g. "[lte][agw] Changeset" or "[orc8r][docker] ..."
-->

## Summary

- `last_reported_time` on subscriber state is updated to float, to fix NMS dependency discrepancy (Date object expects float, due to this it's showing incorrect date)
- Fixing incorrect module naming for cpe_monitoring object
- Updating pipelined.yml to `10.1.0.1` to match mtr0 iface config
- Reformatting code on monitord module

## Test Plan

- Tested locally on AGW, and local Orc8r with eNB and UE to check ICMP monitoring reaches target and reports state correctly

```
(python) vagrant@magma-dev:~/magma/lte/gateway$ sudo journalctl -fu magma@monitord
-- Logs begin at Tue 2020-12-01 07:42:04 UTC. --
Dec 01 08:54:18 magma-dev monitord[13382]: ERROR:root:GetSubscribers Error for StatusCode.UNAVAILABLE! Connect Failed
Dec 01 08:54:18 magma-dev monitord[13382]: WARNING:root:No subscribers/ping targets found
Dec 01 08:55:31 magma-dev monitord[13382]: INFO:root:001010000000082:192.168.128.84 => 179.451ms
Dec 01 08:56:27 magma-dev monitord[13382]: INFO:root:Replicating state for 001010000000082: {"last_reported_time": 1606812931.343585, "latency_ms": 179.451}
Dec 01 08:56:45 magma-dev monitord[13382]: INFO:root:001010000000082:192.168.128.84 => 127.01ms
Dec 01 08:57:27 magma-dev monitord[13382]: INFO:root:Replicating state for 001010000000082: {"last_reported_time": 1606813005.555548, "latency_ms": 127.01}
Dec 01 08:57:58 magma-dev monitord[13382]: INFO:root:001010000000082:192.168.128.84 => 244.653ms
Dec 01 08:58:27 magma-dev monitord[13382]: INFO:root:Replicating state for 001010000000082: {"last_reported_time": 1606813078.78969, "latency_ms": 244.653}
Dec 01 08:59:12 magma-dev monitord[13382]: INFO:root:001010000000082:192.168.128.84 => 172.52ms
Dec 01 08:59:27 magma-dev monitord[13382]: INFO:root:Replicating state for 001010000000082: {"last_reported_time": 1606813152.992067, "latency_ms": 172.52}
Dec 01 09:00:27 magma-dev monitord[13382]: INFO:root:001010000000082:192.168.128.84 => 249.099ms
Dec 01 09:00:27 magma-dev monitord[13382]: INFO:root:Replicating state for 001010000000082: {"last_reported_time": 1606813227.191562, "latency_ms": 249.099}
Dec 01 09:01:27 magma-dev monitord[13382]: INFO:root:Replicating state for 001010000000082: {"last_reported_time": 1606813227.191562, "latency_ms": 249.099}
Dec 01 09:01:41 magma-dev monitord[13382]: INFO:root:001010000000082:192.168.128.84 => 152.87ms
Dec 01 09:02:28 magma-dev monitord[13382]: INFO:root:Replicating state for 001010000000082: {"last_reported_time": 1606813301.393722, "latency_ms": 152.87}
Dec 01 09:02:54 magma-dev monitord[13382]: INFO:root:001010000000082:192.168.128.84 => 239.474ms
Dec 01 09:03:28 magma-dev monitord[13382]: INFO:root:Replicating state for 001010000000082: {"last_reported_time": 1606813374.732781, "latency_ms": 239.474}
Dec 01 09:04:08 magma-dev monitord[13382]: INFO:root:001010000000082:192.168.128.84 => 183.536ms
Dec 01 09:04:28 magma-dev monitord[13382]: INFO:root:Replicating state for 001010000000082: {"last_reported_time": 1606813448.900638, "latency_ms": 183.536}
```

## Additional Information

- [ ] This change is backwards-breaking