@ardzoht 2020-12-01
> @ardzoht Do we have a unit test to verify this feature ?

We have some basic unit tests under monitord module, it might be a good idea to expand on these, I can look into that in a separate issue.